Split up paste_items test wise.
