+++ opengever/base/tests/test_pasting_allowed.py
@@ -1,0 +1,67 @@
+from ftw.builder import Builder
+from ftw.builder import create
+from ftw.testbrowser import browsing
+from opengever.testing import FunctionalTestCase
+
+
+class TestPastingAllowed(FunctionalTestCase):
+
+    def setUp(self):
+        super(TestPastingAllowed, self).setUp()
+        self.grant('Manager')
+
+    @browsing
+    def test_paste_action_not_displayed_for_contactfolder(self, browser):
+        contactfolder = create(Builder('contactfolder'))
+        contact = create(Builder('contact')
+                         .within(contactfolder))
+
+        paths = ['/'.join(contact.getPhysicalPath())]
+        browser.login().open(contactfolder, {'paths:list': paths},
+                             view='copy_items')
+
+        browser.open(contactfolder)
+        actions = browser.css('#plone-contentmenu-actions li').text
+        self.assertSequenceEqual(['Export as Zip', 'Sharing'], actions)
+
+    @browsing
+    def test_paste_action_not_displayed_for_templatedossier(self, browser):
+        templatedossier = create(Builder('templatedossier'))
+        document = create(Builder('document')
+                          .within(templatedossier))
+
+        paths = ['/'.join(document.getPhysicalPath())]
+        browser.login().open(templatedossier, {'paths:list': paths},
+                             view='copy_items')
+
+        browser.open(templatedossier)
+        actions = browser.css('#plone-contentmenu-actions li').text
+        self.assertSequenceEqual(
+            ['Export as Zip', 'Properties', 'Sharing'], actions)
+
+    @browsing
+    def test_paste_action_not_displayed_for_mails(self, browser):
+        dossier = create(Builder('dossier'))
+        document = create(Builder('document').within(dossier))
+        mail = create(Builder('mail').within(dossier))
+
+        paths = ['/'.join(document.getPhysicalPath())]
+        browser.login().open(dossier, {'paths:list': paths}, view='copy_items')
+
+        browser.open(mail)
+        actions = browser.css('#plone-contentmenu-actions li').text
+        self.assertSequenceEqual(
+            ['Export as Zip', 'Properties', 'save attachments'], actions)
+
+    def test_pasting_not_allowed_if_disallowed_subobject_type(self):
+        repofolder = create(Builder('repository'))
+        dossier = create(Builder('dossier')
+                         .within(repofolder))
+        document = create(Builder('document')
+                          .within(dossier))
+
+        dossier.manage_copyObjects(document.id)
+        pasting_allowed_view = repofolder.restrictedTraverse(
+            'is_pasting_allowed')
+        allowed = pasting_allowed_view()
+        self.assertFalse(allowed)
